go-frp: use initFuncs to add click handlers after writing the DOM
The current challenge is that the "reflect" package doesn't appear
to be supported in GopherJS. I would like to use "reflect" in order
to call the provided "update" function on the current model to
generate a new model and then re-render the whole document with the
new model. The provided "update" function takes the current model
and returns a new model and understands the model's type... so this
all should be possible with "reflect".

Eventually, I would like to include a virtual DOM, check the diffs,
and only re-render the portion of the DOM that has changed.

Baby steps.

Change-Id: Ia6f275356bc71cb12dd5b4db083e45fc6f6f0e48